Add service/timer for removing old wal-e backups
This will add service alongside the base-backup service and timer that
will use wal-e to delete the oldest base-backups after making new
base-backups each sunday night.

The number 5 is picked a bit at random, it doesn't seem we run out of
disk that often.

There might be a better way to trigger this than a timer, but I am not
that experienced with systemd services.

Attempts to fix #138